Add Discount gRPC service with SQLite support
Implemented a new service for the Discount gRPC project in `docker-compose.override.yml` and `docker-compose.yml`. This service utilizes SQLite as a file-based database, eliminating the need for an external database container. Configured environment variables for development, HTTP, and HTTPS ports, along with a connection string for the SQLite database. Included volume mounts for user secrets and HTTPS certificates, exposing ports 6002 and 6062 for HTTP and HTTPS, respectively.