Add endpoint to retrieve orders by customer ID
Introduces a new `GetOrdersByCustomer` endpoint in the
`GetOrdersByCustomer.cs` file. This includes a new
`GetOrdersByCustomerResponse` record for response data
and implements the `ICarterModule` interface. The endpoint
maps a GET request to `/orders/customer/{customerId}`,
sending a `GetOrdersByCustomerQuery` and returning the
result as a 200 OK response. It also defines possible
responses for 400 Bad Request and 404 Not Found.